pyproject.toml
[tool.poetry]
name = "vabot"
version = "0.1.0"
description = "VA Bot with ChatGPT API then connected with Telegram"
authors = ["Feri Lukmansyah"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.6.1"
python-dotenv = "^1.0.0"
flask = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
